packages/api_bridge/requirements.txt
fastapi[standard]>=0.110.0
pydantic>=2.7.0
uvicorn>=0.29.0
websockets>=12.0
python-dotenv>=1.0.1
